@@ -1,17 +1,19 @@
-import React from "react"
-import ReactDOM from "react-dom"
-import App from "./App"
-import "./styles/tailwind.css"
-import * as serviceWorker from "./serviceWorker"
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App';
+import './styles/tailwind.css';
+import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-   <React.StrictMode>
-      <App />
-   </React.StrictMode>,
-   document.getElementById("root")
-)
+const root = ReactDOM.createRoot(
+	document.getElementById('root') as HTMLElement
+);
+root.render(
+	<React.StrictMode>
+		<App />
+	</React.StrictMode>
+);
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
