@@ -1,157 +1,172 @@
-import React, { useState, useEffect, memo } from "react"
-import { Link } from "react-router-dom"
+import React, {
+	useState,
+	memo,
+	Dispatch,
+	SetStateAction,
+	useCallback,
+	useEffect,
+} from 'react';
+import { Link } from 'react-router-dom';
 
-import Icon from "./Icon"
-import { IconName } from "@fortawesome/fontawesome-svg-core"
+import Icon from './Icon';
+import { IconName } from '@fortawesome/fontawesome-svg-core';
+import { SectionRef } from '../helpers/types';
+import sections from '../data/sectionData';
 
-type ButtonClickEvent = React.MouseEvent<HTMLButtonElement, MouseEvent>
+type ButtonClickEvent = React.MouseEvent<HTMLButtonElement, MouseEvent>;
 
-interface INavProps {
-   refs: any
-   screenRef: any
+interface NavProps {
+	screenRef: SectionRef;
+	sectionRefs: SectionRef[];
+	view: string;
+	setView: Dispatch<SetStateAction<string>>;
 }
 
-interface INavButtonProps {
-   ref?: React.Ref<HTMLDivElement>
-   icon: IconName
-   label: string
-   color: string
-   isSection: boolean
-   onClick: (e: ButtonClickEvent) => void
+interface NavButtonProps {
+	ref?: React.RefObject<HTMLDivElement>;
+	icon: IconName;
+	label: string;
+	color: string;
+	isSection: boolean;
+	onClick: (e: ButtonClickEvent) => void;
 }
 
-interface INav {
-   label: string
-   icon: IconName
-   color: string
-}
-
-const NavButton: React.FC<INavButtonProps> = memo(
-   React.forwardRef(({ isSection, icon, label, color, onClick }, ref: any) => (
-      <button
-         ref={ref}
-         onClick={onClick}
-         className={`w-full md:w-auto p-4 md:mx-2 md:py-1 cursor-pointer text-sm text-right md:text-center font-bold uppercase border-b-2 
-      ${isSection ? `text-${color} border-${color}` : "border-transparent"}
+const NavButton: React.FC<NavButtonProps> = memo(
+	React.forwardRef(({ isSection, icon, label, color, onClick }, ref: any) => (
+		<button
+			ref={ref}
+			onClick={onClick}
+			className={`w-full md:w-auto p-4 md:mx-2 md:py-1 cursor-pointer text-sm text-right md:text-center font-bold uppercase border-b-2 
+      ${isSection ? `text-${color} border-${color}` : 'border-transparent'}
       hover:border-${color} hover:text-${color} transition-all duration-200`}
-      >
-         <Icon name={icon} size="lg" /> {label}
-      </button>
-   ))
-)
+		>
+			<Icon name={icon} size="lg" /> {label}
+		</button>
+	))
+);
 
-const Nav: React.FC<INavProps> = ({ refs, screenRef }) => {
-   const navs: INav[] = [
-      { label: "intro", icon: "hand-heart", color: "profile-400" },
-      { label: "about", icon: "id-badge", color: "yellow-500" },
-      { label: "skill", icon: "brackets-curly", color: "purple-500" },
-      { label: "experience", icon: "user-chart", color: "green-500" },
-      { label: "project", icon: "tasks", color: "red-500" },
-   ]
+const Nav: React.FC<NavProps> = ({ screenRef, sectionRefs }) => {
+	const [showNav, setNav] = useState<boolean>(false);
+	const [active, setActive] = useState(0);
 
-   const [showNav, setNav] = useState<boolean>(false)
+	const gotoSection = useCallback(
+		(i: number) => () => {
+			if (sectionRefs[i]?.current) {
+				sectionRefs[i]?.current?.scrollIntoView!({ behavior: 'smooth' });
+			}
+		},
+		[sectionRefs]
+	);
 
-   // ===============================================================================
-   const [active, setActive] = useState(0)
+	useEffect(() => {
+		const handle = () => {
+			console.log('in');
+			let currentSectionId = active;
+			for (let i = 0; i < sectionRefs.length; i++) {
+				const iRef = sectionRefs[i];
+				if (iRef?.current) {
+					console.log('in', i);
+					const section = iRef.current;
+					if (!section || !(section instanceof Element)) continue;
+					if (section.getBoundingClientRect().top - 91 < 0) {
+						currentSectionId = i;
+						continue;
+					}
+				}
+				break;
+			}
+			setActive(currentSectionId);
+		};
 
-   const handle = () => {
-      let currentSectionId = active
-      for (let i = 0; i < refs.length; i++) {
-         const iRef: React.MutableRefObject<any> = refs[i]
-         if (iRef) {
-            const section = iRef.current
-            if (!section || !(section instanceof Element)) continue
-            if (section.getBoundingClientRect().top - 61 < 0) {
-               currentSectionId = i
-               continue
-            }
-         }
-         break
-      }
-      setActive(currentSectionId)
-   }
+		const myRef = screenRef;
+		if (myRef?.current) {
+			const screen = myRef.current;
+			screen.addEventListener('scroll', handle);
+			return () => {
+				screen.removeEventListener('scroll', handle);
+			};
+		}
+	}, [active, screenRef, sectionRefs]);
 
-   useEffect(() => {
-      const myRef = screenRef
-      const screen: { addEventListener: any; removeEventListener: any } = myRef.current
-      screen.addEventListener("scroll", handle)
-      return () => {
-         screen.removeEventListener("scroll", handle)
-      }
-   })
+	return (
+		<div className="w-full md:w-5/6 bg-white flex h-24 px-4 md:px-20 md:mx-auto justify-between items-center border-b border-gray-200">
+			<h3 className="whitespace-no-wrap font-bold text-sm md:text-base">
+				<code className="text-yellow-500 font-bold">{'{'}</code> made with{' '}
+				<Icon name="heart" className="text-red-500" alt="love" /> and{' '}
+				<Icon
+					name="react"
+					prefix="fab"
+					className="text-profile-400"
+					alt="react"
+				/>{' '}
+				+{' '}
+				<span
+					className="pl-2 pr-1 font-bold text-white bg-profile cursor-default rounded"
+					title="typescript"
+				>
+					ts
+				</span>{' '}
+				+{' '}
+				<span className="text-sm md:text-base" title="tailwindcss">
+					tailwindcss
+				</span>{' '}
+				<code className="text-yellow-500 font-bold">{'}'}</code>
+			</h3>
+			<button
+				onClick={() => setNav((old) => !old)}
+				onBlur={() => setTimeout(() => setNav(false), 200)}
+				className="block md:hidden"
+			>
+				<Icon name="bars" size="lg" />
+			</button>
+			<div
+				className={`${
+					!showNav && 'hidden'
+				} absolute bg-white w-full top-0 left-0 mt-16 md:mt-0 md:left-auto md:relative z-50 border-b border-gray-300`}
+			>
+				<div className="flex flex-col justify-end items-end z-50">
+					{sections.map((nav, i) => (
+						<NavButton
+							ref={sectionRefs[i]}
+							key={nav.label}
+							isSection={active === i}
+							onClick={gotoSection(i)}
+							{...nav}
+						/>
+					))}
+					<Link
+						to="/files/M_DARY_NUR_RABBANI.pdf"
+						target="_blank"
+						download
+						className="w-full no-underline p-4 cursor-pointer text-sm text-right font-bold text-multicolor"
+					>
+						<Icon name="file-download" size="lg" /> My VC
+					</Link>
+				</div>
+			</div>
+			<div className="hidden md:flex w-full top-0 left-0 mt-16 md:mt-0 md:left-auto md:relative z-50">
+				<div className="w-full flex flex-row justify-end items-center z-50">
+					{sections.map((nav, i) => (
+						<NavButton
+							ref={sectionRefs[i]}
+							key={nav.label}
+							isSection={active === i}
+							onClick={gotoSection(i)}
+							{...nav}
+						/>
+					))}
+					<Link
+						to="/files/M_DARY_NUR_RABBANI.pdf"
+						target="_blank"
+						download
+						className="no-underline p-4 cursor-pointer text-sm text-right font-bold text-multicolor"
+					>
+						<Icon name="file-download" size="lg" /> My VC
+					</Link>
+				</div>
+			</div>
+		</div>
+	);
+};
 
-   // ===============================================================================
-
-   return (
-      <div className="w-full flex h-16 px-4 md:mx-auto justify-between items-center border-b border-gray-200">
-         <h3 className="whitespace-no-wrap font-bold text-sm md:text-base">
-            <code className="text-yellow-500 font-bold">{"{"}</code> made with{" "}
-            <Icon name="heart" className="text-red-500" alt="love" /> and{" "}
-            <Icon name="react" prefix="fab" className="text-profile-400" alt="react" /> +{" "}
-            <span className="pl-2 pr-1 font-bold text-white bg-profile cursor-default rounded" title="typescript">
-               ts
-            </span>{" "}
-            +{" "}
-            <span className="text-sm md:text-base" title="tailwindcss">
-               tailwindcss
-            </span>{" "}
-            <code className="text-yellow-500 font-bold">{"}"}</code>
-         </h3>
-         <button
-            onClick={() => setNav(old => !old)}
-            onBlur={() => setTimeout(() => setNav(false), 200)}
-            className="block md:hidden"
-         >
-            <Icon name="bars" size="lg" />
-         </button>
-         <div
-            className={`${
-               !showNav && "hidden"
-            } absolute bg-white w-full top-0 left-0 mt-16 md:mt-0 md:left-auto md:relative z-50 border-b border-gray-300`}
-         >
-            <div className="flex flex-col justify-end items-end z-50">
-               {navs.map((nav, i) => (
-                  <NavButton
-                     ref={refs[i]}
-                     key={nav.label}
-                     isSection={active === i}
-                     onClick={() => refs[i].current.scrollIntoView({ behavior: "smooth" })}
-                     {...nav}
-                  />
-               ))}
-               <Link
-                  to="/files/M_DARY_NUR_RABBANI.pdf"
-                  target="_blank"
-                  download
-                  className="w-full no-underline p-4 cursor-pointer text-sm text-right font-bold text-multicolor"
-               >
-                  <Icon name="file-download" size="lg" /> My VC
-               </Link>
-            </div>
-         </div>
-         <div className="hidden md:flex w-full top-0 left-0 mt-16 md:mt-0 md:left-auto md:relative z-50">
-            <div className="w-full flex flex-row justify-end items-center z-50">
-               {navs.map((nav, i) => (
-                  <NavButton
-                     ref={refs[i]}
-                     key={nav.label}
-                     isSection={active === i}
-                     onClick={() => refs[i].current.scrollIntoView({ behavior: "smooth" })}
-                     {...nav}
-                  />
-               ))}
-               <Link
-                  to="/files/M_DARY_NUR_RABBANI.pdf"
-                  target="_blank"
-                  download
-                  className="no-underline p-4 cursor-pointer text-sm text-right font-bold text-multicolor"
-               >
-                  <Icon name="file-download" size="lg" /> My VC
-               </Link>
-            </div>
-         </div>
-      </div>
-   )
-}
-
-export default memo(Nav)
+export default memo(Nav);
